requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
arrow>=1.2.3
tabulate>=0.9.0
rich>=13.7.0
win10toast>=0.9
python-dotenv>=1.0.1
